apps/api/requirements.txt
# FastAPI + LangChain requirements

# Web framework
fastapi>=0.115
uvicorn[standard]>=0.34

# LangChain
langchain>=0.3
langchain-community>=0.3

# Database
sqlalchemy>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9

# Redis
redis>=5.0

# Auth
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7

# Settings
pydantic-settings>=2.5

# LLM clients
google-genai>=1.0
google-generativeai>=0.8
langchain-google-genai>=2.0

# File extraction
pypdf>=5.0
pillow>=10.0

# Object storage
boto3>=1.35
